requirements.txt
streamlit==1.49.1
PyPDF2==3.0.1
langchain==0.3.27
python-dotenv==1.1.1
faiss-cpu==1.12.0
openai==1.107.2
huggingface-hub==0.34.4
InstructurEmbedding
sentence_transformers